Edition New Basic component added as a Edition advanced stories
