--- raaghu-components/src/rds-comp-edition-new-basic/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./rds-comp-edition-new-basic";
