--- raaghu-components/src/rds-comp-edition-new-basic/rds-comp-edition-new-basic.tsx
@@ -1,68 +1,0 @@
-import React, { useEffect, useState } from "react";
-import { RdsInput, RdsSelectList } from "../rds-elements";
-
-export interface RdsCompEditionNewBasicProps {
-    planList: any[];
-    accountTwoFactorSettings: any;
-    planListLabel?: string;
-    reset?: boolean;
-}
-const RdsCompEditionNewBasic = (props: RdsCompEditionNewBasicProps) => {
-    const [inputReset, setInputReset] = useState(false);
- const [twoFactorData, settwoFactorData] = useState(
-        props.accountTwoFactorSettings
-    );
-
-    const handlerChangeTwoFact = (value: any, key: string) => {
-        settwoFactorData({ ...twoFactorData, [key]: value });
-    };
-
-    useEffect(() => {
-        settwoFactorData(props.accountTwoFactorSettings);
-    }, [props.accountTwoFactorSettings]);
-
-    useEffect(() => {
-        setInputReset(!inputReset);
-    }, [props.reset]);
-    
-    return (
-        <>
-            <form >
-                <div className="row mt-2">
-                    <div className="col-md-6 mb-3">
-                        <div className="form-group">
-                            <RdsInput
-                                label={"Edition Name"}
-                                required={true}
-                                placeholder="Edition Name"
-                                inputType="text"
-                                name="editionName"
-                                reset={inputReset}
-                                dataTestId="edition-name"
-                            ></RdsInput>
-                        </div>
-                    </div>
-                    <div className="col-md-6 mb-3">
-                        <div className="form-group">
-                            <RdsSelectList
-                                id="planLis"
-                                label={props.planListLabel}
-                                isDisabled={false}
-                                isMultiple={false}
-                                selectItems={props.planList}
-                                selectedValue={twoFactorData?.planList}
-                                onChange={(item: any) => {
-                                    handlerChangeTwoFact(item.value, "planList");
-                                }}
-
-                                dataTestId="plan-list"
-                                required={true}
-                            />
-                        </div>
-                    </div>
-                </div>
-            </form>
-        </>
-    );
-};
-export default RdsCompEditionNewBasic;
