+++ raaghu-components/src/rds-comp-edition/rds-comp-edition.stories.tsx
@@ -10,6 +10,10 @@ const meta: Meta = {
     },
     tags: ['autodocs'],
     argTypes: {
+        displayType: {
+            options: ["basic", "advanced"],
+            control: { type: "select" },
+        },
     },
 } satisfies Meta<typeof RdsCompEdition>;
 
@@ -18,6 +22,7 @@ type Story = StoryObj<typeof RdsCompEdition>;
 
 export const Default: Story = {
     args: {
+        displayType: "basic",
         EditionItems: {
             EditionName: "Corporate",
             EditionTitle: "Strong Application for large team",
@@ -28,6 +33,27 @@ export const Default: Story = {
             "Maximum User Count",
             "Test Check feature",
             "Test check feature count 2",
+        ],
+    }
+} satisfies Story;
+
+export const Advanced: Story = {
+    args: {
+        displayType: "advanced",
+        planListLabel : "Plan",
+        planList: [    
+            {
+                "isFree": true,
+                "value": "standard",
+                "option": "Standard",
+                "isSelected": false
+            },
+            {
+                "isFree": false,
+                "value": "advanced",
+                "option": "Advanced",
+                "isSelected": false
+            }
         ],
     }
 } satisfies Story;
