+++ raaghu-components/src/rds-comp-edition/rds-comp-edition.tsx
@@ -17,6 +17,10 @@ export interface RdsCompEditionProps {
     editionName: any;
     reset?: boolean;
     onSaveHandler?: (data: any) => void;
+    planList: any[];
+    accountTwoFactorSettings: any;
+    planListLabel?: string;
+    displayType: "basic" | "advanced";
 }
 const RdsCompEdition = (props: RdsCompEditionProps) => {
     const offCanvasHandler = () => { };
@@ -24,6 +28,7 @@ const RdsCompEdition = (props: RdsCompEditionProps) => {
     const [activeNavTabId, setActiveNavTabId] = useState("0");
     const [showTenantSettings, setShowTenantSettings] = useState(false);
     const [inputReset, setInputReset] = useState(false);
+    const [twoFactorData, settwoFactorData] = useState(props.accountTwoFactorSettings);
     
     useEffect(() => {
         setFormData(props.editionName);
@@ -42,6 +47,19 @@ const RdsCompEdition = (props: RdsCompEditionProps) => {
         { label: "Features", tablink: "#nav-profile", id: 1 },
     ];
 
+    const handlerChangeTwoFact = (value: any, key: string) => {
+        settwoFactorData({ ...twoFactorData, [key]: value });
+    };
+
+    useEffect(() => {
+        settwoFactorData(props.accountTwoFactorSettings);
+    }, [props.accountTwoFactorSettings]);
+
+    useEffect(() => {
+        setInputReset(!inputReset);
+    }, [props.reset]);
+    
+
     function emitSaveData(event: any) {
         event.preventDefault();
         props.onSaveHandler && props.onSaveHandler(FormData);
@@ -53,6 +71,8 @@ const RdsCompEdition = (props: RdsCompEditionProps) => {
     }
 
     return (
+        <>
+        {props.displayType == "basic" && (
         <div className="col-md-3 navsm mb-3 featureList ng-star-inserted">
             <div className="card border-undefined">
                 <div className="headerClass">
@@ -208,6 +228,45 @@ const RdsCompEdition = (props: RdsCompEditionProps) => {
                 </div>
             </div>
         </div>
+        )}
+        {props.displayType == "advanced" && (
+            <form >
+            <div className="row mt-2">
+                <div className="col-md-6 mb-3">
+                    <div className="form-group">
+                        <RdsInput
+                            label={"Edition Name"}
+                            required={true}
+                            placeholder="Edition Name"
+                            inputType="text"
+                            name="editionName"
+                            reset={inputReset}
+                            dataTestId="edition-name"
+                        ></RdsInput>
+                    </div>
+                </div>
+                <div className="col-md-6 mb-3">
+                    <div className="form-group">
+                        <RdsSelectList
+                            id="planLis"
+                            label={props.planListLabel}
+                            isDisabled={false}
+                            isMultiple={false}
+                            selectItems={props.planList}
+                            selectedValue={twoFactorData?.planList}
+                            onChange={(item: any) => {
+                                handlerChangeTwoFact(item.value, "planList");
+                            }}
+
+                            dataTestId="plan-list"
+                            required={true}
+                        />
+                    </div>
+                </div>
+            </div>
+        </form>
+        )}
+        </>
     );
 };
